Add confirmation message to install scripts
Some install script didn't have any confirmation message.
Those script would proceed without user consent.
Now the user is able to stop the script before anything happens.

+++ install-nvm.sh
@@ -1,5 +1,16 @@
 #!/bin/bash
 
+echo "This script installs nvm"
+echo ""
+echo "Are you sure you want to continue?"
+read -p "Press [y/Y] to confirm: " answer
+echo ""
+
+if [ "$answer" != "y" ] && [ "$answer" != "Y" ]; then
+	echo "canceled"
+	exit
+fi
+
 curl -o- https://raw.githubusercontent.com/nvm-sh/nvm/v0.35.2/install.sh | bash
 echo "open a new terminal and type 'command -v nvm' to test the installation"
